requirements.txt
# ==============================================================================
# PROJET ERP DISTRIBUTION - Dépendances Python globales
# ==============================================================================
# Installation : pip install -r requirements.txt
# ==============================================================================

# Base de données et configuration
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Analyse de données
pandas==2.1.4
numpy==1.24.4

# Machine Learning et Data Mining
scikit-learn==1.3.2

# Visualisation
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Templates et rapports
jinja2==3.1.2

# API et requêtes HTTP
requests>=2.31

# IA et LLM
google-genai>=1.0.0

# ==============================================================================
# Notes d'installation
# ==============================================================================
# 1. Créer l'environnement virtuel :
#    python -m venv venv
#
# 2. Activer l'environnement :
#    Windows : venv\Scripts\activate
#    Linux/Mac : source venv/bin/activate
#
# 3. Installer les dépendances :
#    pip install -r requirements.txt
#
# 4. Copier et configurer l'environnement :
#    copy .env.example .env
#    # Éditer .env avec vos configurations
# ==============================================================================
